Another iteration in getting a subset templated version compiled and linked.

+++ cantera/branches/jacobians/include/cantera/thermo/GeneralSpeciesThermo.h
@@ -208,9 +208,9 @@ public:
 
 #ifdef H298MODIFY_CAPABILITY
 
-    virtual doublereal reportOneHf298(int k) const;
+    virtual doublereal reportOneHf298(size_t k) const;
 
-    virtual void modifyOneHf298(const int k, const doublereal Hf298New);
+    virtual void modifyOneHf298(const size_t k, const doublereal Hf298New);
 
 #endif
