+++ cantera/branches/jacobians/src/thermo/NasaThermo.h
@@ -426,7 +426,7 @@ public:
     }
 
 #ifdef H298MODIFY_CAPABILITY
-    virtual doublereal reportOneHf298(const int k) const {
+    virtual doublereal reportOneHf298(const size_t k) const {
 
         int grp = m_group_map[k];
         int pos = m_posInGroup_map[k];
@@ -444,7 +444,7 @@ public:
         return h;
     }
 
-    virtual void modifyOneHf298(const int k, const doublereal Hf298New) {
+    virtual void modifyOneHf298(const size_t k, const doublereal Hf298New) {
         int grp = m_group_map[k];
         int pos = m_posInGroup_map[k];
         std::vector<NasaPoly1> &mlg = m_low[grp-1];
